requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
matplotlib
tensorflow
tensorflow.keras
wordcloud
joblib
scikit-learn
comet-ml
dvc
